Move timer to platform
