+++ src/platform/timing.cpp
@@ -1,4 +1,4 @@
-#include <timing.h>
+#include <platform/timing.h>
 
 namespace timing {
