+++ src/platform/timing.h
